Refactor HR Zones chart, clean up ActivitySummary
[frontend] moved HR Zones graph away from ActivitySummaryComponent
[frontend] removed duplicated translation
[frontend] reverted changes on ActivitySummaryComponent since chart was removed
[frontend] moved chart to ActivityMandAbovePillsComponent
[frontend] small adjustments to BarChart
