+++ frontend/app/src/i18n/us/components/activities/activityMandAbovePillsComponent.json
@@ -4,11 +4,13 @@
     "labelPillWorkoutSets": "Sets",
     "labelGraph": "Activity data graphs",
     "labelGraphHR": "Heart rate",
+    "labelHRZones": "Heart rate zones",
     "labelGraphPower": "Power",
     "labelGraphCadence": "Cadence",
     "labelGraphElevation": "Elevation",
     "labelGraphVelocity": "Speed",
     "labelGraphPace": "Pace",
+    "labelGraphHRZone": "Zone",
     "labelDownsampling": "Data downsampled to ~200 points",
     "errorMessageProcessingActivityStreams": "Error processing activity streams",
     "labelGraphStrokeRate": "Stroke rate"
