+++ frontend/app/src/i18n/us/components/activities/activitySummaryComponent.json
@@ -10,18 +10,12 @@
     "labelVirtual": "(Virtual) ",
     "activityDistance": "Distance",
     "activityTime": "Time",
-    "activityTotalTime": "Total Time",
-    "activityMovingTime": "Moving Time",
-    "activityElevationGain": "Elevation Gain",
     "activityPace": "Pace",
-    "activityAvgPower": "Avg Power",
-    "activityAvgSpeed": "Avg Speed",
-    "activityHR": "Heart Rate",
     "activityAvgHR": "Avg HR",
-    "activityMaxHR": "Max HR",
-    "activityHRZone": "Zone",
-    "activityEleGain": "Elevation gain",
-    "activityEleLoss": "Elevation loss",
+    "activityAvgPower": "Avg power",
+    "activityAvgSpeed": "Avg speed",
+    "activityEleGain": "Ele gain",
+    "activityEleLoss": "Ele loss",
     "activityCalories": "Calories",
     "activityNoData": "No data",
     "errorFetchingUserById": "Error fetching user by id",
